Returning byte instead of int.

+++ src/nim/x86.nim
@@ -1,6 +1,6 @@
 include private/x86.nim
 
-proc inc*(operand: reg16, buf: var ptr byte): int =
+proc inc*(operand: reg16, buf: var ptr byte): byte =
   var offset = 0
   cast[ptr byte](buf)[] = 0x66 + prefix_adder(operand)
   buf += 1
@@ -8,7 +8,7 @@ proc inc*(operand: reg16, buf: var ptr byte): int =
   buf += 1
   return offset
 
-proc inc*(operand: reg32, buf: var ptr byte): int =
+proc inc*(operand: reg32, buf: var ptr byte): byte =
   var offset = 0
   if (operand > 7):
     cast[ptr byte](buf)[] = 65
@@ -17,7 +17,7 @@ proc inc*(operand: reg32, buf: var ptr byte): int =
   buf += 1
   return offset
 
-proc dec*(operand: reg16, buf: var ptr byte): int =
+proc dec*(operand: reg16, buf: var ptr byte): byte =
   var offset = 0
   cast[ptr byte](buf)[] = 0x66 + prefix_adder(operand)
   buf += 1
@@ -25,7 +25,7 @@ proc dec*(operand: reg16, buf: var ptr byte): int =
   buf += 1
   return offset
 
-proc dec*(operand: reg32, buf: var ptr byte): int =
+proc dec*(operand: reg32, buf: var ptr byte): byte =
   var offset = 0
   if (operand > 7):
     cast[ptr byte](buf)[] = 65
@@ -34,7 +34,7 @@ proc dec*(operand: reg32, buf: var ptr byte): int =
   buf += 1
   return offset
 
-proc push*(operand: reg16, buf: var ptr byte): int =
+proc push*(operand: reg16, buf: var ptr byte): byte =
   var offset = 0
   cast[ptr byte](buf)[] = 0x66 + prefix_adder(operand)
   buf += 1
@@ -42,7 +42,7 @@ proc push*(operand: reg16, buf: var ptr byte): int =
   buf += 1
   return offset
 
-proc push*(operand: reg32, buf: var ptr byte): int =
+proc push*(operand: reg32, buf: var ptr byte): byte =
   var offset = 0
   if (operand > 7):
     cast[ptr byte](buf)[] = 65
@@ -51,7 +51,7 @@ proc push*(operand: reg32, buf: var ptr byte): int =
   buf += 1
   return offset
 
-proc pop*(operand: reg16, buf: var ptr byte): int =
+proc pop*(operand: reg16, buf: var ptr byte): byte =
   var offset = 0
   cast[ptr byte](buf)[] = 0x66 + prefix_adder(operand)
   buf += 1
@@ -59,7 +59,7 @@ proc pop*(operand: reg16, buf: var ptr byte): int =
   buf += 1
   return offset
 
-proc pop*(operand: reg32, buf: var ptr byte): int =
+proc pop*(operand: reg32, buf: var ptr byte): byte =
   var offset = 0
   if (operand > 7):
     cast[ptr byte](buf)[] = 65
@@ -68,7 +68,7 @@ proc pop*(operand: reg32, buf: var ptr byte): int =
   buf += 1
   return offset
 
-proc pop*(operand: reg64, buf: var ptr byte): int =
+proc pop*(operand: reg64, buf: var ptr byte): byte =
   var offset = 0
   cast[ptr byte](buf)[] = 0x48 + prefix_adder(operand)
   buf += 1
@@ -76,17 +76,17 @@ proc pop*(operand: reg64, buf: var ptr byte): int =
   buf += 1
   return offset
 
-proc pushf*(buf: var ptr byte): int =
+proc pushf*(buf: var ptr byte): byte =
   cast[ptr byte](buf)[] = 156
   buf += 1
   return 1
 
-proc popf*(buf: var ptr byte): int =
+proc popf*(buf: var ptr byte): byte =
   cast[ptr byte](buf)[] = 157
   buf += 1
   return 1
 
-proc ret*(buf: var ptr byte): int =
+proc ret*(buf: var ptr byte): byte =
   cast[ptr byte](buf)[] = 195
   buf += 1
   return 1
